Move WebSocketCommunicator<JSON> to apps/querier
and remove serde_json dependency from lib
