+++ apps/querier/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2018"
 
 [dependencies]
 rustyweb = { path = "../../lib" }
+serde_json = "^1.0"
