+++ lib/Cargo.toml
@@ -7,4 +7,3 @@ edition = "2018"
 [dependencies]
 rust-crypto = "^0.2"
 base64 = "^0.10"
-serde_json = "^1.0"
