+++ lib/src/lib.rs
@@ -1,4 +1,3 @@
 pub mod http;
 pub mod web;
 pub mod parser;
-pub mod websocket;
